@bjorn3 2020-11-07
Locally I get

```
test cache_messages::caching_large_output ... test cache_messages::caching_large_output has been running for over 60 seconds
```

I think this is a deadlock due to message printing and job running no longer happening on different threads, so job running fills the message channel and waits, but the message won't be processed, as the thread responsible for it is blocked.