@bjorn3 2020-11-07
I worked around it for now by removing the backpressure. A better solution would be to immediately process the messages.