@bjorn3 2020-11-07
> A better solution would be to immediately process the messages

Done