@bjorn3 2020-11-09
> Thanks for this! Would it be possible to split apart the thread-spawn change from the error allocation change?

Sure, #8844.