@alexcrichton 2020-11-11
@bors: r+ 

Ok this looks good to me as-is and after talking a brief amount more in the Cargo team meeting today the conclusion is that if the main thread takes a long time copying output from a big rustc invocation that's not a huge issue, so I'm gonna approve.

Thanks again @bjorn3!